+++ src/onegov.town/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Correctly sort the the pages even if the title has changed.
+  [href]
+
 - Limits the news list on the homepage to two entries.
   [href]